move pingAddress to cmd package - replace post keys flow to sequential
