--- eth2util/keystore/keystore_internal_test.go
@@ -1,31 +1,0 @@
-// Copyright © 2022 Obol Labs Inc.
-//
-// This program is free software: you can redistribute it and/or modify it
-// under the terms of the GNU General Public License as published by the Free
-// Software Foundation, either version 3 of the License, or (at your option)
-// any later version.
-//
-// This program is distributed in the hope that it will be useful, but WITHOUT
-// ANY WARRANTY; without even the implied warranty of  MERCHANTABILITY or
-// FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
-// more details.
-//
-// You should have received a copy of the GNU General Public License along with
-// this program.  If not, see <http://www.gnu.org/licenses/>.
-
-package keystore
-
-import (
-	"net/http/httptest"
-	"testing"
-
-	"github.com/stretchr/testify/require"
-)
-
-func TestPingAddress(t *testing.T) {
-	srv := httptest.NewServer(nil)
-	defer srv.Close()
-
-	require.NoError(t, pingAddress(srv.URL))
-	require.Error(t, pingAddress("1.1.1.1"))
-}
